==> foreign_import_analysis/22_fao_transfers_by_ecoregion.R <==
# Convert incoming land transfers from each country to incoming land transfers from ecoregions
# QDR / Virtualland / 17 Feb 2021

# Modified 18 Feb 2021: do separately for the 20 scenarios

# Needed: incoming virtual cropland and pastureland countries by country
# Needed: total area of cropland and pastureland in each ecoregion in each country

library(data.table)
library(Rutilitybelt)

# Read the tabulated counts of crop and pasture by country x TNC
count_cropd <- fread(file.path(spatial_output_path, 'global_count_cropdominance.csv'))
count_pasture <- fread(file.path(spatial_output_path, 'global_count_pasture.csv'))

# Extract the useful information from the cropland summaries
# Use crop dominance. Conservative estimate by excluding the minor fragment pixels. 
# Should not be sensitive to this because it is only for relative areas.
# In crop dominance, classes 1-7 are cropland.
count_cropd <- count_cropd[, crop_sum := rowSums(count_cropd[,c(as.character(1:7))], na.rm = TRUE)]

# Read the land transfers by FAO
fao_vlt <- fread(file.path(intermediate_output_path, 'fao_VLT.csv'))

# Remove zero VLT countries
fao_vlt[, VLT_sum := rowSums(.SD, na.rm = TRUE), .SDcols = patterns('VLT_')]
fao_vlt <- fao_vlt[VLT_sum > 0]

# Join the VLT country list with the intersected country x TNC polygon
# There is no code to join on so we have to do it by name

dput(setdiff(fao_vlt$country_name, count_cropd$NAME_LONG)) # About 20 don't match.

# Change the names of the FAO VLT dataframe to match the country names in countrymap.
names_to_match <- c("Bolivia (Plurinational State of)", "Cabo Verde", "China, mainland", 
                    "Congo", "Gambia", "China, Hong Kong SAR", "Iran (Islamic Republic of)", 
                    "C\xf4te d'Ivoire", "Republic of Moldova", "North Macedonia", 
                    "Czechia", "Sao Tome and Principe", "Eswatini", "Syrian Arab Republic", 
                    "China, Taiwan Province of", "United Republic of Tanzania", "United Kingdom of Great Britain and Northern Ireland", 
                    "Venezuela (Bolivarian Republic of)")
names_corrected <- c("Bolivia", "Republic of Cabo Verde", "China", 
                     "Democratic Republic of the Congo", "The Gambia", "Hong Kong", "Iran",
                     "Côte d'Ivoire", "Moldova", "Macedonia",
                     "Czech Republic", "São Tomé and Principe", "eSwatini", "Syria",
                     "Taiwan", "Tanzania", "United Kingdom", "Venezuela")

for (i in 1:length(names_to_match)) {
  fao_vlt$country_name[fao_vlt$country_name %in% names_to_match[i]] <- names_corrected[i]
}

# Extract needed columns from the crop dominance and pasture count data frames, and join the two.
count_cropd <- count_cropd[, .(ECO_CODE, ECO_NAME, WWF_REALM, NAME_LONG, ISO_A3, REGION_UN, SUBREGION, area, crop_sum)]
count_pasture <- count_pasture[, .(ECO_CODE, ECO_NAME, WWF_REALM, NAME_LONG, ISO_A3, REGION_UN, SUBREGION, area, sum)]
setnames(count_cropd, old = 'crop_sum', new = 'crop_area')
setnames(count_pasture, old = 'sum', new = 'pasture_area')

country_tnc_data <- count_pasture[count_cropd, on = .(ECO_CODE, ECO_NAME, WWF_REALM, NAME_LONG, ISO_A3, REGION_UN, SUBREGION, area)]

# Determine, for each country, the relative proportion of cropland and pastureland in each ecoregion.
# This will be used to get the relative proportion of virtual cropland and pastureland transfers 
# from each ecoregion x country combination.

replace_na_dt(country_tnc_data, cols = c('crop_area', 'pasture_area'))
country_tnc_data[, c('crop_proportion', 'pasture_proportion') := .(crop_area / sum(crop_area), pasture_area / sum(pasture_area)), by = NAME_LONG]

# Join country_tnc_data with the virtual land transfers, proportionally split by area
# Cartesian join because of the 20 scenarios
foreign_vlt_eco <- country_tnc_data[fao_vlt, on = c('NAME_LONG' = 'country_name'), allow.cartesian = TRUE]
foreign_vlt_eco[, VLT_annual_region := VLT_annual * crop_proportion]
foreign_vlt_eco[, VLT_mixed_region := VLT_mixed * crop_proportion]
foreign_vlt_eco[, VLT_permanent_region := VLT_permanent * crop_proportion]
foreign_vlt_eco[, VLT_pasture_region := VLT_pasture * pasture_proportion]

# Separate scenario column into two
foreign_vlt_eco <- tidyr::separate(foreign_vlt_eco, scenario, into = c('D', 'scenario_diet', 'W', 'scenario_waste'), sep = '_')
foreign_vlt_eco[, c('D','W') := NULL]

# Sum the land transfers across ecoregions
foreign_vlt_eco_sum <- foreign_vlt_eco[, lapply(.SD, sum),
                                       by = .(scenario_diet, scenario_waste, ECO_CODE, ECO_NAME),
                                       .SDcols = patterns('region|pasture_area|crop_area')]


fwrite(foreign_vlt_eco, file.path(final_output_path, 'foreign_VLT_by_country_x_TNC.csv'))
fwrite(foreign_vlt_eco_sum, file.path(final_output_path, 'foreign_VLT_by_TNC.csv'))
